+++ src/staging/package.json
@@ -41,6 +41,7 @@
     "react-typography": "^0.16.19",
     "sass": "^1.37.5",
     "sharp": "^0.28.3",
+    "simple-react-lightbox": "^3.6.9-0",
     "typography": "^0.16.19",
     "typography-theme-kirkham": "^0.16.19"
   },
